Creates userModel.js and userRoutes.js
userModel.js outlines the mongodb collection of users with username, password, email, and isAdmin. All are required and the username must be unique.

userRoutes.js has a series of api calls that will be used for querying the database. The first being register which requires username and email and sets isAdmin to default to false. it also takes the password and generates a hash and then stores the password to make sure no one can see it. Login is after and it will check if login was completed or not. There needs to be some other response right now it just sends it to the console. update password updates and stores the new hash and check admin returns the true or false value on whether a user is or isn't an admin.